less6 is completed
All tasks are fully ready
